Remove context from StateMachine
There is no need to keep a context for handlers explicitly as a part of
StateMachine. If you need to keep state between handler calls, just create a
struct and use bound methods as handlers. Then you can access the struct from
each and every handler and it can server as your context.

Signed-off-by: Ondrej Kupka <ondra.cap@gmail.com>